# vendor: implement --versioned-dirs