--- a/src/etc/man/cargo-vendor.1
+++ b/src/etc/man/cargo-vendor.1
@@ -57,10 +57,10 @@ Don\(cqt delete the "vendor" directory when vendoring, but rather keep all
 existing contents of the vendor directory.
 .RE
 .sp
-\fB\-\-explicit\-version\fP
+\fB\-\-versioned\-dirs\fP
 .RS 4
 Normally versions are only added to disambiguate multiple versions of the same package.
-This option causes all directories in the "vendor" directory to be explicitly versioned,
+This option causes all directories in the "vendor" directory to be versioned,
 which makes it easier to track the history of vendored packages over time, and can help
 with the performance of re-vendoring when only a subset of the packages have changed.
 .RE