--- a/src/etc/man/cargo-vendor.1
+++ b/src/etc/man/cargo-vendor.1
@@ -54,7 +54,7 @@ vendored and synced to the output.
 \fB\-\-no\-delete\fP
 .RS 4
 Don\(cqt delete the "vendor" directory when vendoring, but rather keep all
-existing contents of the vendor directory
+existing contents of the vendor directory.
 .RE
 .sp
 \fB\-\-explicit\-version\fP
@@ -68,7 +68,7 @@ with the performance of re-vendoring when only a subset of the packages have cha
 \fB\-\-respect\-source\-config\fP
 .RS 4
 Instead of ignoring \fB[source]\fP configuration by default in \fB.cargo/config\fP
-read it and use it when downloading crates from crates.io, for example
+read it and use it when downloading crates from crates.io, for example.
 .RE
 .SS "Manifest Options"
 .sp