@alexcrichton 2019-11-26
I think it'd also be good to write up some documentation to explain this feature and why it might be used.