@jsgf 2019-11-27
Test and doc added. LMK if you think there's more to be tested/documented.