@jsgf 2019-12-10
@joshtriplett @ehuss Documentation updated, and I think all comments addressed.